Added the menu class and made an if statement after Login() to check if user is admin or not. Also wrote out a different Login() with authenticate()
