+++ RebelAllianceBank/Classes/Bank.cs
@@ -4,13 +4,93 @@ namespace RebelAllianceBank.Classes
 {
     public class Bank
     {
-        List<IUser> users = new List<IUser>() { new Admin("FullAccessLogin", "02492512", "Admin", "Adminson") };
+        List<IUser> users;
         IUser? currentUser;
 
         public void Run()
         {
+            FileHandler fh = new FileHandler();
+            users = new List<IUser>(fh.ReadUser());
             Login();
+            if (currentUser is Admin)
+            {
+                Menu.AdminMenu(users);
+            }
+            else
+            {
+                Menu.CustomerMenu(currentUser);
+            }
         }
+
+        //public void Login()
+        //{
+        //    while (true)
+        //    {
+        //        Console.Clear();
+        //        Console.WriteLine("Välkommen till Rebel Alliance Bank. Vänligen ange ditt personnummer.");
+        //        Console.Write("Personnummer: ");
+        //        string? usernameInput = Console.ReadLine();
+
+        //        if (string.IsNullOrWhiteSpace(usernameInput))
+        //        {
+        //            Console.WriteLine("Personnummer får inte vara tomt. Försök igen.");
+        //            Console.ReadKey();
+        //            continue;
+        //        }
+
+        //        currentUser = users.FirstOrDefault(user => user.PersonalNum == usernameInput);
+        //        if (currentUser != null)
+        //        {
+        //            if (currentUser.LoginLock)
+        //            {
+        //                Console.WriteLine("Användaren är låst. Kontakta administratör.\n" +
+        //                                  "Tryck på enter för att fortsätta");
+        //                while (Console.ReadKey(true).Key != ConsoleKey.Enter) { }
+        //            }
+        //            else if (Authenticate(currentUser))
+        //            {
+        //                break;
+        //            }
+        //        }
+        //        else
+        //        {
+        //            Console.WriteLine("Användarnamnet hittades inte. Tryck på enter för att försöka igen.");
+        //            while (Console.ReadKey(true).Key != ConsoleKey.Enter) { }
+        //        }
+        //    }
+        //}
+        //public bool Authenticate(IUser user)
+        //{
+        //    int tries = 3;
+        //    while (tries > 0)
+        //    {
+        //        Console.Clear();
+        //        Console.WriteLine($"God dag {user.Surname}. Vänligen skriv ditt lösenord. Du har {tries} försök kvar.");
+        //        string? passwordInput = Console.ReadLine();
+
+        //        if (user.Password == passwordInput)
+        //        {
+        //            return true;
+        //        }
+        //        else
+        //        {
+        //            tries--;
+        //            Console.WriteLine($"Felaktigt lösenord. Du har {tries} försök kvar.");
+        //            if (tries == 0)
+        //            {
+        //                Console.WriteLine("Användaren är nu låst. Kontakta administratör!\n" +
+        //                    "Tryck på enter för att återgå");
+        //                while (Console.ReadKey(true).Key != ConsoleKey.Enter) { }
+        //                user.LoginLock = true;
+        //                return false;
+        //            }
+        //            Console.WriteLine($"Tryck på enter för att försöka igen.");
+        //            while (Console.ReadKey(true).Key != ConsoleKey.Enter) { }
+        //        }
+        //    }
+        //    return false;
+        //}
+
 
         /// <summary>
         /// Method that runs the login system and loops until successful login was done.
@@ -20,14 +100,15 @@ namespace RebelAllianceBank.Classes
             while (true)
             {
                 Console.Clear();
-                Console.WriteLine("Välkommen till Rebel Alliance Bank. Vänligen skriv ditt personnummer.");
+                Console.WriteLine("Välkommen till Rebel Alliance Bank. Vänligen ange ditt personnummer.");
+                Console.Write("Personnummer: ");
                 string? usernameInput = Console.ReadLine();
                 bool correctUser = false;
                 bool correctPass = false;
                 bool userLocked = false;
                 int tries = 0;
 
-                //Checks if inputted username is a valid username. Also checks if that user is locked from logging in.
+                //Checks if inputted username is a valid username.Also checks if that user is locked from logging in.
                 foreach (var user in users)
                 {
                     if (user.PersonalNum.Equals(usernameInput) && user.LoginLock == true)
@@ -49,7 +130,7 @@ namespace RebelAllianceBank.Classes
                 if (correctUser)
                 {
                     tries = 0;
-                    
+
                     //Loops until correct password is inputted or if wrong password is inputted 3 times.
                     while (true)
                     {
@@ -86,218 +167,16 @@ namespace RebelAllianceBank.Classes
                         }
                     }
                 }
-                else if(!userLocked)
+                else if (!userLocked)
                 {
                     Console.WriteLine("Det finns ingen användare med det användarnamnet. Tryck på valfri " +
                         "tangent för att gå tillbaka och försöka igen.");
                     Console.ReadKey();
                 }
 
-                if(correctPass)
+                if (correctPass)
                 {
                     break;
-                }
-            }
-        }
-
-        public static void AdminMenu()
-        {
-            bool runAdminMenu = true;
-
-            while (runAdminMenu)
-            {
-                Console.Clear();
-
-                Console.Write("ADMIN\n" +
-                              "[1] Skapa användare\n" +
-                              "[2] Ändra växelkurs\n" +
-                              "[3] Lås upp användarkonto???\n" +
-                              "[4] Logga ut\n" +
-                              "\n" +
-                              "Menyval: ");
-
-                string choice = Console.ReadLine();
-
-                switch (choice)
-                {
-                    case "1":
-                        Console.WriteLine("Skapa användare");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "2":
-                        Console.WriteLine("Ändra växelkurs");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "3":
-                        Console.WriteLine("Lås upp användarkonto???");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "4":
-                        runAdminMenu = false;
-                        break;
-                    default:
-                        Console.Clear();
-                        Console.WriteLine("Felaktig input! Tryck enter och försök igen!");
-                        Console.ReadKey();
-                        break;
-                }
-            }
-        }
-
-        public static void CustomerMenu()
-        {
-            bool runCustomerMenu = true;
-
-            while (runCustomerMenu)
-            {
-                Console.Clear();
-                Console.Write($"Välkommen {"John Doe"}!\n" +
-                              $"\n" +
-                              $"HUVUDMENY:\n" +
-                              $"[1] Konton\n" +
-                              $"[2] Betala/Överföra\n" +
-                              $"[3] Lån\n" +
-                              $"[4] Logga ut\n" +
-                              $"\n" +
-                              $"Menyval: ");
-
-                string choice = Console.ReadLine();
-
-                switch (choice)
-                {
-                    case "1":
-                        CustomerMenuAccounts();
-                        break;
-                    case "2":
-                        CustomerMenuTransaction();
-                        break;
-                    case "3":
-                        CustomerMenuLoan();
-                        break;
-                    case "4":
-                        runCustomerMenu = false;
-                        break;
-                    default:
-                        Console.Clear();
-                        Console.WriteLine("Felaktig input! Tryck enter och försök igen!");
-                        Console.ReadKey();
-                        break;
-                }
-            }
-        }
-
-        private static void CustomerMenuAccounts()
-        {
-            bool runCustomerMenuAccounts = true;
-
-            while (runCustomerMenuAccounts)
-            {
-                Console.Clear();
-
-                Console.Write("KONTON:\n" +
-                              "[1] Se över mina konton\n" +
-                              "[2] Öppna nytt konto\n" +
-                              "[3] Återgå till huvudmenyn\n" +
-                              "\n" +
-                              "Menyval: ");
-
-                string choice = Console.ReadLine();
-
-                switch (choice)
-                {
-                    case "1":
-                        Console.WriteLine("Se över mina konton");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "2":
-                        Console.WriteLine("Öppna nytt konto");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "3":
-                        runCustomerMenuAccounts = false;
-                        break;
-                    default:
-                        Console.Clear();
-                        Console.WriteLine("Felaktig input! Tryck enter och försök igen!");
-                        Console.ReadKey();
-                        break;
-                }
-            }
-        }
-        private static void CustomerMenuTransaction()
-        {
-            bool runCustomerMenuTransaction = true;
-
-            while (runCustomerMenuTransaction)
-            {
-                Console.Clear();
-
-                Console.Write("BETALA/ÖVERFÖRA:\n" +
-                              "[1] Ny överföring\n" +
-                              "[2] Ny betalning\n" +
-                              "[3] Återgå till huvudmenyn\n" +
-                              "\n" +
-                              "Menyval: ");
-
-                string choice = Console.ReadLine();
-
-                switch (choice)
-                {
-                    case "1":
-                        Console.WriteLine("Ny överföring");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "2":
-                        Console.WriteLine("Ny betalning");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "3":
-                        runCustomerMenuTransaction = false;
-                        break;
-                    default:
-                        Console.Clear();
-                        Console.WriteLine("Felaktig input! Tryck enter och försök igen!");
-                        Console.ReadKey();
-                        break;
-                }
-            }
-        }
-
-        private static void CustomerMenuLoan()
-        {
-            bool runCustomerMenuLoan = true;
-
-            while (runCustomerMenuLoan)
-            {
-                Console.Clear();
-
-                Console.Write("LÅN:\n" +
-                              "[1] Mina lån\n" +
-                              "[2] Ansök om nytt lån\n" +
-                              "[3] Återgå till huvudmenyn\n" +
-                              "\n" +
-                              "Menyval: ");
-
-                string choice = Console.ReadLine();
-
-                switch (choice)
-                {
-                    case "1":
-                        Console.WriteLine("Mina lån");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "2":
-                        Console.WriteLine("Ansök om nytt lån");
-                        Console.ReadKey(); //Ta ev bort sen när det finns en metod
-                        break;
-                    case "3":
-                        runCustomerMenuLoan = false;
-                        break;
-                    default:
-                        Console.Clear();
-                        Console.WriteLine("Felaktig input! Tryck enter och försök igen!");
-                        Console.ReadKey();
-                        break;
                 }
             }
         }
@@ -423,65 +302,65 @@ namespace RebelAllianceBank.Classes
         /// <summary>
         /// Method that runs function to unlock a locked user.
         /// </summary>
-        public void UnlockUser()
-        {
-            while (true)
-            {
-                Console.WriteLine("Skriv användarnamnet av den användare du vill låsa upp. Skriv exit om du vill gå" +
-                    " tillbaka till menyn");
-                string usernameInput = Console.ReadLine();
-                bool correctInput = false;
-                bool notLockedUser = false;
+        //public void UnlockUser()
+        //{
+        //    while (true)
+        //    {
+        //        Console.WriteLine("Skriv användarnamnet av den användare du vill låsa upp. Skriv exit om du vill gå" +
+        //            " tillbaka till menyn");
+        //        string usernameInput = Console.ReadLine();
+        //        bool correctInput = false;
+        //        bool notLockedUser = false;
 
-                //Checks if user wants to exit from function and breaks loop if exit is inputted.
-                if(usernameInput.ToUpper().Equals("EXIT"))
-                {
-                    break;
-                }
+        //        //Checks if user wants to exit from function and breaks loop if exit is inputted.
+        //        if (usernameInput.ToUpper().Equals("EXIT"))
+        //        {
+        //            break;
+        //        }
 
-                foreach (var user in users)
-                {
-                    //Checks if users username is the inputted username and checks if that user is locked. If not tells
-                    // current user that that useraccount isn't locked and waits for a key press to go back to input.
-                    if (user.PersonalNum.Equals(usernameInput) && user.LoginLock == false)
-                    {
-                        Console.WriteLine("Användaren är inte låst. Kolla om du skrivit rätt användarnamn. Tryck på valfri rangent" +
-                            " för att gå vidare.");
-                        correctInput = true;
-                        notLockedUser = true;
-                        Console.ReadKey();
-                        Console.Clear();
-                        break;
-                    }
-                    //Checks if users username is the inputted username and unlocks that useraccount if it is and it is
-                    //locked.
-                    else if (user.PersonalNum.Equals(usernameInput))
-                    {
-                        user.LoginLock = false;
-                        Console.WriteLine("Användaren har låsts upp. Tryck på valfri tangent för att gå vidare.");
-                        correctInput = true;
-                        Console.ReadKey();
-                        Console.Clear();
-                        break;
-                    }
-                }
+        //        foreach (var user in users)
+        //        {
+        //            //Checks if users username is the inputted username and checks if that user is locked. If not tells
+        //            // current user that that useraccount isn't locked and waits for a key press to go back to input.
+        //            if (user.PersonalNum.Equals(usernameInput) && user.LoginLock == false)
+        //            {
+        //                Console.WriteLine("Användaren är inte låst. Kolla om du skrivit rätt användarnamn. Tryck på valfri rangent" +
+        //                    " för att gå vidare.");
+        //                correctInput = true;
+        //                notLockedUser = true;
+        //                Console.ReadKey();
+        //                Console.Clear();
+        //                break;
+        //            }
+        //            //Checks if users username is the inputted username and unlocks that useraccount if it is and it is
+        //            //locked.
+        //            else if (user.PersonalNum.Equals(usernameInput))
+        //            {
+        //                user.LoginLock = false;
+        //                Console.WriteLine("Användaren har låsts upp. Tryck på valfri tangent för att gå vidare.");
+        //                correctInput = true;
+        //                Console.ReadKey();
+        //                Console.Clear();
+        //                break;
+        //            }
+        //        }
 
-                //Continues the loop if a correct username was inputted but that useraccount wasn't locked.
-                if(correctInput && notLockedUser)
-                {
-                    continue;
-                }
-                else if(correctInput)
-                {
-                    break;
-                }
-                else
-                {
-                    Console.WriteLine("Det finns ingen användare med det användarnamnet.tryck på " +
-                        "valfri tangent för att gå tillbaka och försök igen.");
-                }
-            }
+        //        //Continues the loop if a correct username was inputted but that useraccount wasn't locked.
+        //        if (correctInput && notLockedUser)
+        //        {
+        //            continue;
+        //        }
+        //        else if (correctInput)
+        //        {
+        //            break;
+        //        }
+        //        else
+        //        {
+        //            Console.WriteLine("Det finns ingen användare med det användarnamnet.tryck på " +
+        //                "valfri tangent för att gå tillbaka och försök igen.");
+        //        }
+        //    }
 
-        }
+        //}
     }
 }
