+++ RebelAllianceBank/Classes/FileHandler.cs
@@ -58,7 +58,7 @@ namespace RebelAllianceBank.Classes
         /// <returns>An <see cref="IUser"/> object or null if the row is invalid.</returns>
         public IUser StoredUser(string[] row)
         {
-            switch (row[3])
+            switch (row[5])
             {
                 case "true":
                     return new Admin
@@ -74,7 +74,7 @@ namespace RebelAllianceBank.Classes
                     return new Customer
                     {
                         ID = Convert.ToInt16(row[0]),
-                        PersonalNum = row[2],
+                        PersonalNum = row[1],
                         Password = row[2],
                         Surname = row[3],
                         Forename = row[4],
