+++ RebelAllianceBank/users.txt
@@ -1,2 +1,2 @@
-001-9101010101-12345-Gunnar-Larsson
-
+001-9101010101-12345-Gunnar-Larsson-false
+001-9201010101-12345-Gunnar-Larsson-true
